ci: Publish alpha versions with alpha tag
We need to publish alpha versions with an alpha tag so that it is not
automatically installed if the package version isn't pinned to a
specific version in the package.json file of consuming projects. By
default, if we do not specify a tag, it will be tagged as `latest`
which would allow any consumer to inadvertently install it.

See https://docs.npmjs.com/adding-dist-tags-to-packages